Separate remote loopback runtime bridge ownership
The browser SSH loopback bridge had grown inside the alias policy type, which made the policy owner also responsible for a large WebKit runtime script. Split the runtime injection script into its own helper while keeping the alias constants as the single source of truth used by both Swift routing and injected JavaScript.

Constraint: Preserve the existing remote loopback alias invariant and avoid direct xcodebuild per project build policy.

Rejected: Leave the bridge on RemoteLoopbackProxyAlias | keeps alias policy and WebKit runtime script ownership mixed.

Confidence: high

Scope-risk: narrow

Directive: Keep loopback host policy in RemoteLoopbackProxyAlias and browser runtime API shims in RemoteLoopbackRuntimeBridge.

Tested: Node bridge syntax and rewrite harness for localhost 3000/8000, subdomain localhost, ws, wss, and https; plutil lint of project.pbxproj; git diff --check.

Not-tested: Local XCTest and app build per workflow; PR CI and final reload.sh launch will verify.
